+++ test/support/site/controller/page_controller.ex
@@ -1,5 +1,6 @@
 defmodule Test.Site.PageController do
   use Phoenix.Controller, layouts: [html: {Test.Site.Layout, :basic}]
+  use Phoenix.VerifiedRoutes, endpoint: Test.Site.Endpoint, router: Test.Site.Router
   import Phoenix.Component, only: [to_form: 2]
 
   def show(conn, _params), do: render(conn, :show)
@@ -12,12 +13,13 @@ defmodule Test.Site.PageController do
     |> render(:form)
   end
 
-  def submit(conn, %{"form" => params}) do
-    changeset = changeset(params)
+  def submit(conn, %{"form" => form_params} = params) do
+    changeset = changeset(form_params)
+    send(self(), {:page_controller, :submit, params})
 
     case Ecto.Changeset.apply_action(changeset, :insert) do
       {:ok, _} ->
-        redirect(conn, to: :show)
+        redirect(conn, to: ~p"/pages/show")
 
       {:error, changeset} ->
         conn